--- test/Jvmhs/Data/AnnotationSpec.hs
@@ -1,12 +1,0 @@
-{-# OPTIONS_GHC -fno-warn-orphans #-}
-{-# LANGUAGE TypeSynonymInstances #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE FlexibleInstances #-}
-module Jvmhs.Data.AnnotationSpec where
-
-import           SpecHelper
-
--- import           Jvmhs.Data.Annotation
-
-spec :: Spec
-spec = return ()
